Move Juniper test to grammar test
